requirements.txt
flask
flask_cors
gunicorn
google-generativeai
python-dotenv
soundfile
numpy
pydantic
langchain-community
sentence-transformers
torch
chromadb